requirements.txt
appdirs==1.4.3
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
breezy==3.0.2
cached-property==1.5.1
catkin-pkg==0.4.16
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
dulwich==0.19.15
empy==3.3.2
entrypoints==0.3
evdev==1.4.0
fastimport==0.9.8
filelock==3.0.12
genmsg==0.5.14
genpy==0.6.9
gpg===1.13.1-unknown
httplib2==0.14.0
i3ipc==2.1.1
idna==2.8
importlib-metadata==1.5.0
jsonschema==3.2.0
keyboard==0.13.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
more-itertools==4.2.0
MouseInfo==0.1.3
mss==6.1.0
netifaces==0.10.4
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
PyAutoGUI==0.9.52
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGetWindow==0.0.9
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pynput==1.7.3
pyparsing==2.4.6
pyperclip==1.8.2
PyQt5==5.14.1
PyRect==0.1.4
pyRFC3339==1.1
pyrsistent==0.15.5
PyScreeze==0.1.26
pytesseract==0.3.7
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-xlib==0.23
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rofication==1.2.2
roman==2.0.0
rosdep==0.18.0
rosdistro==0.8.0
rosinstall==0.7.8
rospkg==1.2.3
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
ssh-import-id==5.10
system-service==0.3
systemd-python==234
texttable==1.6.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vboxapi==1.0
vcstools==0.1.42
virtualenv==20.0.17
wadllib==1.3.3
websocket-client==0.53.0
wrapt==1.11.2
wstool==0.1.18
xkit==0.0.0
zipp==1.0.0
